+++ metadata.rb
@@ -1,10 +1,10 @@
-name             'iis-nativex'
-maintainer       'NativeX'
-maintainer_email 'adrian.herrera@nativex.com'
+name             'iis-simplyadrian'
+maintainer       'simplyadrian'
+maintainer_email 'simplyadrian@gmail.com'
 license          'All rights reserved'
-description      'Installs/Configures iis-nativex'
+description      'Installs/Configures iis-simplyadrian'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'
 
-depends           "windows"
-depends           "iis"
+depends  'windows'
+depends  'iis'