+++ recipes/create-app-pool.rb
@@ -1,13 +1,13 @@
 #
-# Cookbook Name:: iis-nativex
+# Cookbook Name:: iis-simplyadrian
 # Recipe:: create-app-pool
 #
-# Copyright 2015, NativeX
+# Copyright 2015, simplyadrian
 #
 # All rights reserved - Do Not Redistribute
 #
 
-node['iis-nativex']['enabled_pools'].each do |pool_cfg|
+node['iis-simplyadrian']['enabled_pools'].each do |pool_cfg|
   iis_pool pool_cfg['pool_name'] do
     runtime_version pool_cfg['runtime_version']
     pipeline_mode pool_cfg['pipeline_mode']