+++ recipes/create-config-sites.rb
@@ -1,13 +1,13 @@
 #
-# Cookbook Name:: iis-nativex
+# Cookbook Name:: iis-simplyadrian
 # Recipe:: create-config-sites
 #
-# Copyright 2015, NativeX
+# Copyright 2015, simplyadrian
 #
 # All rights reserved - Do Not Redistribute
 #
 
-Chef::Resource::IisSite.send(:include, IisNativex::Helper)
+Chef::Resource::IisSite.send(:include, IisSimplyadrian::Helper)
 
 # stop and delete the default site
 iis_site 'Default Web Site' do
@@ -15,7 +15,7 @@ iis_site 'Default Web Site' do
 end
 
 #create physical locations
-node['iis-nativex']['enabled_sites'].each do |mkdir|
+node['iis-simplyadrian']['enabled_sites'].each do |mkdir|
   directory "#{node['iis']['docroot']}/#{mkdir['site_name']}" do
     recursive true
     action :create
@@ -24,7 +24,7 @@ end
 
 # create and start a new site that maps to
 # the physical location C:\inetpub\wwwroot\#{site_name}
-node['iis-nativex']['enabled_sites'].each do |site_cfg|
+node['iis-simplyadrian']['enabled_sites'].each do |site_cfg|
   iis_site site_cfg[:site_name] do
     protocol site_cfg[:protocol]
     port site_cfg[:port]